#3 Add TestHelper for common test code. Remove SparkTestSuite as its irrelevant.
